pyproject.toml
[project]
name = "dragon-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioboto3>=15.5.0",
    "dotenv>=0.9.9",
    "fastapi>=0.132.0",
    "httpx[http2]>=0.28.1",
    "imagehash>=4.3.2",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "tenacity>=9.1.4",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]
